backend/scripts/requirements.txt
fugashi==1.5.0
unidic-lite==1.0.8
mysql-connector-python==8.2.0
